+++ tests/ActiveFilter.test.ts
@@ -1,58 +1,45 @@
 import ActiveFilter from '../src/ActiveFilter';
-import defaultOptions from '../src/defaultOptions';
 
 describe('ActiveFilter', () => {
   describe('#constructor', () => {
     it('should create an instance with default filter set to "all"', () => {
-      expect(new ActiveFilter(defaultOptions).filter).toEqual('all');
+      expect(new ActiveFilter('all').get()).toEqual('all');
     });
   });
 
-  describe('#get filter', () => {
+  describe('#get', () => {
     it('returns the value of the filter', () => {
-      const activeFilter = new ActiveFilter(defaultOptions);
-      expect(activeFilter.filter).toEqual('all');
+      const activeFilter = new ActiveFilter('all');
+      expect(activeFilter.get()).toEqual('all');
     });
   });
 
   describe('#set', () => {
     it('set the value of the filter to passed value', () => {
-      const activeFilter = new ActiveFilter(defaultOptions);
+      const activeFilter = new ActiveFilter('all');
       activeFilter.set('5');
-      expect(activeFilter.filter).toEqual('5');
+      expect(activeFilter.get()).toEqual('5');
     });
   });
 
   describe('#toggle', () => {
     it('toggle the value of the filter passed in', () => {
-      const activeFilter1 = new ActiveFilter({
-        ...defaultOptions,
-        filter: '5',
-      });
+      const activeFilter1 = new ActiveFilter('5');
       activeFilter1.toggle('5');
 
-      const activeFilter2 = new ActiveFilter({
-        ...defaultOptions,
-        filter: ['4', '8'],
-      });
+      const activeFilter2 = new ActiveFilter(['4', '8']);
       activeFilter2.toggle('8');
 
-      const activeFilter3 = new ActiveFilter({
-        ...defaultOptions,
-        filter: ['4', '8'],
-      });
+      const activeFilter3 = new ActiveFilter(['4', '8']);
       activeFilter3.toggle('9');
 
-      const activeFilter4 = new ActiveFilter({
-        ...defaultOptions,
-        filter: 'all',
-      });
+      const activeFilter4 = new ActiveFilter('all');
       activeFilter4.toggle('2');
 
-      expect(activeFilter1.filter).toEqual('all');
-      expect(activeFilter2.filter).toEqual('4');
-      expect(activeFilter3.filter).toEqual(['4', '8', '9']);
-      expect(activeFilter4.filter).toEqual('2');
+      expect(activeFilter1.get()).toEqual('all');
+      expect(activeFilter2.get()).toEqual('4');
+      expect(activeFilter3.get()).toEqual(['4', '8', '9']);
+      expect(activeFilter4.get()).toEqual('2');
     });
   });
 });